requirements.txt
Pillow==6.2.1
numpy